refactor: split discovery handling out of update_resources_loaded.go
Move updateAPIResourceDiscovery, handleAPIResourceDiscoveryError, and
updateDiscoveryCacheLoaded into a new update_resources_discovery.go.
Extract the bookmark-replay and deferred-session-restore blocks out of
updateAPIResourceDiscovery into their own methods to bring it under
the gocyclo cap.
